backend-old/requirements.txt
alabaster==0.7.16
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
docutils==0.21.2
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
requests==2.32.3
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
zipp==3.21.0
